requirements.txt
selenium==3.141.0
allure-behave
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pyTelegramBotAPI
telethon
gspread
antigate